Revert "pretty owo"
This reverts commit a4fb29377f8713a4d6372762bb85284e3fcc67fc.

# Conflicts:
#	dao/karafile.ts
#	utils/files.ts

+++ utils/previews.ts
@@ -5,22 +5,15 @@ import { KaraList } from '../types/kara';
 import { resolvedPathPreviews, resolvedPathRepos } from './config';
 import { createThumbnail, extractAlbumArt } from './ffmpeg';
 import { resolveFileInDirs } from './files';
-import logger, { profile } from './logger';
+import logger, {profile} from './logger';
 
 let creatingThumbnails = false;
 
-export async function createImagePreviews(
-	karas: KaraList,
-	thumbnailType?: 'single' | 'full',
-	width = 600
-) {
+export async function createImagePreviews(karas: KaraList, thumbnailType?: 'single' | 'full', width = 600) {
 	if (karas.content.length === 0) return;
 	thumbnailType = thumbnailType || 'full'; // default
 	if (creatingThumbnails) {
-		logger.warn(
-			'Creating previews in progress, please wait a moment and try again later',
-			{ service: 'Previews' }
-		);
+		logger.warn('Creating previews in progress, please wait a moment and try again later', {service: 'Previews'});
 		return;
 	}
 	creatingThumbnails = true;
@@ -36,8 +29,7 @@ export async function createImagePreviews(
 		const fileParts = file.split('.');
 		if (mediaMap.has(fileParts[0])) {
 			// Compare mediasizes. If mediasize is different, remove file
-			if (mediaMap.get(fileParts[0]) !== +fileParts[1])
-				fs.unlink(resolve(resolvedPathPreviews(), file));
+			if (mediaMap.get(fileParts[0]) !== +fileParts[1]) fs.unlink(resolve(resolvedPathPreviews(), file));
 		}
 	});
 	profile('removePreviews');
@@ -47,23 +39,13 @@ export async function createImagePreviews(
 		const kara = karas.content[index];
 		const counter = +index + 1;
 		try {
-			if (
-				!previewSet.has(
-					`${kara.kid}.${kara.mediasize}.25${width > 600 ? '.hd' : ''}.jpg`
-				)
-			) {
+			if (!previewSet.has(`${kara.kid}.${kara.mediasize}.25${width > 600 ? '.hd':''}.jpg`)) {
 				if (!kara.mediafile.endsWith('.mp3')) {
-					logger.debug(
-						`Creating thumbnails for ${kara.mediafile} (${counter}/${karas.content.length})`,
-						{ service: 'Previews' }
-					);
+					logger.debug(`Creating thumbnails for ${kara.mediafile} (${counter}/${karas.content.length})`, {service: 'Previews'});
 					let mediaPath: string[];
 					try {
-						mediaPath = await resolveFileInDirs(
-							kara.mediafile,
-							resolvedPathRepos('Medias')
-						);
-					} catch (err) {
+						mediaPath = await resolveFileInDirs(kara.mediafile, resolvedPathRepos('Medias'));
+					} catch(err) {
 						continue;
 					}
 					const creates = [
@@ -74,55 +56,47 @@ export async function createImagePreviews(
 							kara.mediasize,
 							kara.kid,
 							width
-						),
-					];
+						)];
 					if (thumbnailType === 'full') {
-						creates.push(
-							createThumbnail(
-								mediaPath[0],
-								33,
-								kara.duration,
-								kara.mediasize,
-								kara.kid,
-								width
-							)
-						);
-						creates.push(
-							createThumbnail(
-								mediaPath[0],
-								50,
-								kara.duration,
-								kara.mediasize,
-								kara.kid,
-								width
-							)
-						);
+						creates.push(createThumbnail(
+							mediaPath[0],
+							33,
+							kara.duration,
+							kara.mediasize,
+							kara.kid,
+							width
+						));
+						creates.push(createThumbnail(
+							mediaPath[0],
+							50,
+							kara.duration,
+							kara.mediasize,
+							kara.kid,
+							width
+						));
 					}
 					await Promise.all(creates);
 				} else {
-					logger.debug(
-						`Creating thumbnail for ${kara.mediafile} (${counter}/${karas.content.length})`,
-						{ service: 'Previews' }
-					);
+					logger.debug(`Creating thumbnail for ${kara.mediafile} (${counter}/${karas.content.length})`, {service: 'Previews'});
 					let mediaPath: string[];
 					try {
-						mediaPath = await resolveFileInDirs(
-							kara.mediafile,
-							resolvedPathRepos('Medias')
-						);
-					} catch (err) {
+						mediaPath = await resolveFileInDirs(kara.mediafile, resolvedPathRepos('Medias'));
+					} catch(err) {
 						continue;
 					}
-					await extractAlbumArt(mediaPath[0], kara.mediasize, kara.kid, width);
+					await extractAlbumArt(
+						mediaPath[0],
+						kara.mediasize,
+						kara.kid,
+						width
+					);
 				}
 			}
 		} catch (error) {
-			logger.debug(
-				`Error when creating thumbnail for ${kara.mediafile}: ${error}`,
-				{ service: 'Previews' }
-			);
+			logger.debug(`Error when creating thumbnail for ${kara.mediafile}: ${error}`, {service: 'Previews'});
 		}
 	}
 	profile('createPreviews');
 	creatingThumbnails = false;
 }
+
